# Add October CacheRoute plugin to speed up page loading