--- a/plugins/serenitynow/cacheroute/Plugin.php
+++ b/plugins/serenitynow/cacheroute/Plugin.php
@@ -1,0 +1,14 @@
+<?php namespace SerenityNow\Cacheroute;
+
+use System\Classes\PluginBase;
+
+class Plugin extends PluginBase
+{
+    public function boot()
+    {
+        //add RouteCacheMiddleware as a global middleware to intercept all routes
+        $this->app['Illuminate\Contracts\Http\Kernel']
+             ->pushMiddleware('SerenityNow\CacheRoute\Classes\RouteCacheMiddleware');
+    }
+    
+}