--- a/plugins/serenitynow/cacheroute/controllers/cacheroutes/_list_toolbar.htm
+++ b/plugins/serenitynow/cacheroute/controllers/cacheroutes/_list_toolbar.htm
@@ -1,0 +1,25 @@
+<div data-control="toolbar">
+        <a href="<?= Backend::url('serenitynow/cacheroute/cacheroutes/create') ?>" class="btn btn-primary oc-icon-plus"><?= e(trans('backend::lang.form.create')) ?></a>
+        <a href="<?= Backend::url('serenitynow/cacheroute/cacheroutes/reorder') ?>" class="btn btn-primary oc-icon-sitemap"><?= e(trans('backend::lang.reorder.default_title')) ?></a>
+        <button
+        class="btn btn-default oc-icon-trash-o"
+        disabled="disabled"
+        onclick="$(this).data('request-data', {
+            checked: $('.control-list').listWidget('getChecked')
+        })"
+        data-request="onDelete"
+        data-request-confirm="<?= e(trans('backend::lang.list.delete_selected_confirm')) ?>"
+        data-trigger-action="enable"
+        data-trigger=".control-list input[type=checkbox]"
+        data-trigger-condition="checked"
+        data-request-success="$(this).prop('disabled', true)"
+        data-stripe-load-indicator>
+        <?= e(trans('backend::lang.list.delete_selected')) ?>
+    </button>
+        <button
+        class="btn btn-warning oc-icon-trash-o"
+        data-request="onClear"
+        data-stripe-load-indicator>
+            Clear ALL cached content
+        </button>
+</div>