--- a/plugins/serenitynow/cacheroute/controllers/cacheroutes/_reorder_toolbar.htm
+++ b/plugins/serenitynow/cacheroute/controllers/cacheroutes/_reorder_toolbar.htm
@@ -1,0 +1,3 @@
+<div data-control="toolbar">
+    <a href="<?= Backend::url('serenitynow/cacheroute/cacheroutes') ?>" class="btn btn-primary oc-icon-caret-left"><?= e(trans('backend::lang.form.return_to_list')) ?></a>
+</div>