--- a/plugins/serenitynow/cacheroute/controllers/cacheroutes/config_reorder.yaml
+++ b/plugins/serenitynow/cacheroute/controllers/cacheroutes/config_reorder.yaml
@@ -1,0 +1,5 @@
+title: CacheRoutes
+modelClass: SerenityNow\Cacheroute\Models\CacheRoute
+nameFrom: route_pattern
+toolbar:
+    buttons: reorder_toolbar