--- a/plugins/serenitynow/cacheroute/controllers/cacheroutes/index.htm
+++ b/plugins/serenitynow/cacheroute/controllers/cacheroutes/index.htm
@@ -1,0 +1,1 @@
+<?= $this->listRender() ?>