--- a/plugins/serenitynow/cacheroute/controllers/cacheroutes/reorder.htm
+++ b/plugins/serenitynow/cacheroute/controllers/cacheroutes/reorder.htm
@@ -1,0 +1,8 @@
+<?php Block::put('breadcrumb') ?>
+    <ul>
+        <li><a href="<?= Backend::url('serenitynow/cacheroute/cacheroutes') ?>">CacheRoutes</a></li>
+        <li><?= e($this->pageTitle) ?></li>
+    </ul>
+<?php Block::endPut() ?>
+
+<?= $this->reorderRender() ?>