--- a/plugins/serenitynow/cacheroute/lang/en/lang.php
+++ b/plugins/serenitynow/cacheroute/lang/en/lang.php
@@ -1,0 +1,11 @@
+<?php return [
+    'plugin' => [
+        'name' => 'CacheRoute',
+        'description' => 'Speeds up your application by caching output associated with specified routes',
+        'route_pattern' => 'Route Pattern',
+        'cache_ttl' => 'TTL',
+        'cache_ttl_comment' => 'TTL In Minutes',
+        'manage_cache_route' => 'Manage CacheRoute',
+        'sort_order' => 'Sort Order',
+    ],
+];