This provides a significant speedup by caching the CMS' pages. This way, the database no longer needs to be queried and templates no longer need to be rendered to display pages.

The October CacheRoute plugin source is here: https://github.com/mnshankar/oc-cacheroute-plugin
This pull request includes commit https://github.com/mnshankar/oc-cacheroute-plugin/commit/f0d91ef04f7c600407fa2f955835dfe35a20d3f8.

There's no need to bother setting up the plugin for development.

Recommended configuration for production **(the order matters!)**:

| Route pattern | TTL (Time To Live)      |
|---------------|-------------------------|
| `backend/*`   | 0 minutes               |
| `download`    | 0 minutes               |
| `/`           | 5 minutes               |
| `article/*`   | 5 minutes               |
| `news*`       | 5 minutes               |
| `devblog*`    | 5 minutes               |
| `*`           | 525600 minutes (1 year) |

Static pages are cached for a long duration, pages featuring dynamic (blog) content are cached for a short duration, and backend pages are not cached at all. The download per-OS redirect is also not cached either as that causes issues; only the OS-specific download pages are cached.

For `news*` and `devblog*`, it's important not to add a slash as we want to also match the first page of the list (whose URL is just `/news` or `/devblog`).

Make sure to run `php artisan plugin:refresh serenitynow.cacheroute` in the October folder after deploying the plugin.

Remember to clear the cache after a deployment by running `php artisan cache:clear` in the October folder. Also, after configuring the route patterns, clear the cache using the October backend menu option.

## How much of a speedup does this provide?

### Without the plugin

```
❯ hey http://localhost:8000/community

Summary:
  Total:	10.7954 secs
  Slowest:	2.7755 secs
  Fastest:	0.0528 secs
  Average:	2.3684 secs
  Requests/sec:	18.5263


Response time histogram:
  0.053 [1]	|
  0.325 [5]	|■
  0.597 [5]	|■
  0.870 [5]	|■
  1.142 [6]	|■■
  1.414 [5]	|■
  1.686 [5]	|■
  1.959 [5]	|■
  2.231 [5]	|■
  2.503 [5]	|■
  2.775 [153]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■


Latency distribution:
  10% in 1.0830 secs
  25% in 2.6293 secs
  50% in 2.7064 secs
  75% in 2.7372 secs
  90% in 2.7564 secs
  95% in 2.7656 secs
  99% in 2.7751 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0018 secs, 0.0528 secs, 2.7755 secs
  DNS-lookup:	0.0017 secs, 0.0011 secs, 0.0026 secs
  req write:	0.0000 secs, 0.0000 secs, 0.0001 secs
  resp wait:	2.3656 secs, 0.0500 secs, 2.7729 secs
  resp read:	0.0008 secs, 0.0007 secs, 0.0012 secs

Status code distribution:
  [200]	200 responses
```

### With the plugin

```
❯ hey http://localhost:8000/community

Summary:
  Total:	3.5234 secs
  Slowest:	0.9366 secs
  Fastest:	0.0246 secs
  Average:	0.7651 secs
  Requests/sec:	56.7632


Response time histogram:
  0.025 [1]	|
  0.116 [5]	|■
  0.207 [5]	|■
  0.298 [6]	|■■
  0.389 [5]	|■
  0.481 [5]	|■
  0.572 [6]	|■■
  0.663 [5]	|■
  0.754 [5]	|■
  0.845 [6]	|■■
  0.937 [151]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■


Latency distribution:
  10% in 0.3640 secs
  25% in 0.8551 secs
  50% in 0.8602 secs
  75% in 0.8627 secs
  90% in 0.9174 secs
  95% in 0.9298 secs
  99% in 0.9346 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0025 secs, 0.0246 secs, 0.9366 secs
  DNS-lookup:	0.0022 secs, 0.0000 secs, 0.0036 secs
  req write:	0.0000 secs, 0.0000 secs, 0.0005 secs
  resp wait:	0.7620 secs, 0.0187 secs, 0.9337 secs
  resp read:	0.0005 secs, 0.0001 secs, 0.0040 secs

Status code distribution:
  [200]	200 responses
```