@Calinou 2021-01-11
**Note:** I can't figure out why `php artisan plugin:refresh serenitynow.cacheroute` isn't working in productuion (even though the folder *is* present in `plugins/`). It just says the plugin isn't found…

I disabled the plugin in production for now.